fix: Code generation in README.md
Paste documentation examples to `README.md` in a loop instead of
duplicating the same lines for all of the engines.

Also, changed signature of the `serpapi.Client` to accept three
parameters instead of a dict. This mentally distinguishes arguments
allowed to the client from the `Client#search()` method.

Also, fixed `make all` execution by prepending `python -m` to Python
commands.

Co-authored-by: Dmitriy Zub <dmitriy@serpapi.com>